% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{blockr.dataeditr-package}
\alias{blockr.dataeditr}
\alias{blockr.dataeditr-package}
\title{blockr.dataeditr: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Michael Page \email{mike@cynkra.com}

}
\keyword{internal}
