% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit.R
\name{new_data_edit_block}
\alias{new_data_edit_block}
\title{Edit cells in a data frame}
\usage{
new_data_edit_block(...)
}
\arguments{
\item{...}{Forwarded to \code{\link[blockr.core:new_block]{blockr.core::new_block()}}}
}
\description{
Edit cells in a data frame
}
\examples{
\dontrun{
serve(new_data_edit_block(), data = list(.data = head(mtcars)))
serve(new_data_edit_block(), data = list(.data = data.frame(stat = c("30", "40 (50)"), pattern = c("{mean}", {"{n} ({\%})"}))))
}
}
